requirement.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
selectolax
python-dotenv